backend/requirements_final.txt
deskew==0.10.40
lmdb==1.3.0
natsort==8.1.0
packaging==21.3
pandas==1.4.0
pyspellchecker==0.6.3
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
scipy==1.4.1
six==1.16.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing-extensions==4.0.1
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
efficientnet==1.0.0
Flask==2.1.1
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==2.10.0
idna==3.3
image-classifiers==1.0.0
imageio==2.16.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.1.1
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
segmentation-models @ git+https://github.com/qubvel/segmentation_models.git@94f624b7029deb463c859efbd92fa26f512b52b8
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tifffile==2022.4.8
urllib3==1.26.9
Werkzeug==2.1.1
wincertstore==0.2
wrapt==1.14.0
zipp==3.8.0
